Add id to individual job
The value associated with jobs key is an indexed array of individual job
that is composed of an indexed array of title, salary and location.  Add
id to each job and let it have id, title, salary and location.  Each job
now has a self-reference as $job['id'].

Adding id is the precursor to the database managing the jobs array data.