Add individual job view and route
Build the individual job view to show the job title, salary and location
of the specified job.

Build the route to each job by copying the same array of jobs.  Search
for the job by id and upon finding it, pass it to view.  If no job is
found, return 404.