Use Eloquent model for jobs array from MySQL database
Now the jobs array data is stored in MySQL database.  Use Eloquent model
for the data and delete the same data stored in Job class.

By convention the database table should be named jobs, but that name is
used for another table.  Use job_listings for our database table.

The use of parent::find() inside Job class makes the individual job not
accessible.  Move the check of !$job into the controller temporarily.