Add employer to the job listings
Extract the employer name from the passed-in jobs and add it above the
job listing.  Use block layout to let employer name and job listing flow
vertically.  Add color to text and border to spruce up the listings.

The passed-in jobs are a collection.  Use API to manage the attributes.