Add the route of /jobs/create
The route is added in the routes and it will never be reached.  The
route with wildcard above this, /jobs/{id}, is expanded to /jobs/create
and returns 404.

To make /jobs/create route work in the presence of wildcard routes, let
wildcard routes be parsed last.  In our case, it applies to routes with
/jobs prefix only.