Build the form for submit
Set the form method to POST and action to default route.  Add @csrf
directive to generate token that gets rid of 419 error.

Add the corresponding POST route for create view in routes file.

With all these set up, the form is ready to submit the user input to be
saved in the database if it passes validation.