Use a resource route in lieu of group
Resource routes follow RESTful conventions, we can define a resource
route and apply middleware to it, except for PATCH jobs/{job} route.
Without this exception, PUT will creep in.

For the exception, we can use the usual way for update action.

Routes remain the same as before the changes and route:list confirms it.

$ sail artisan route:list --path=jobs