Use id() in lieu of user()
The method of Auth::id() is a shortcut that directly fetches the id of
the authenticated user.  If no user is logged in, it simply returns
null.

The method of Auth::user() retrieves the authenticated user object and
user()->id returns the value of id property.  If no user is logged in,
Auth::user() returns null, and attempting to access ->id on null will
throw an error (Trying to get property 'id' of non-object).

The use of Auth::id() is safer and more robust.  It handles the case
where no user is logged in gracefully by returning null.  This avoids
potential errors in situations where user authentication might not be
guaranteed.