Enforce policy to authorize 'update' or 'delete' action
The policy system of Laravel provides an explicit and consistent way to
enforce permissions for various actions on a model.  Through the use of
$this->authorize(), the current user must pass the authorization defined
for the update or delete policy on the Job instance.

AuthorizesRequests trait affords authorize() through the 'use' keyword.
By injecting AuthorizesRequests in JobController class, the methods and
properties defined in the trait become part of the class.