Use employer id for adding new job
An employer owns multiple jobs and an authenticated user can authorize
many employers.  The belongsTo relationship involving Job, Employer and
User has replaced the relationship between Job and User.  In alignment
with this change, use employer id when adding a new job and the employer
id is determined by the authenticated user id.

In case of test user, abort and skip adding the new job.