Add controller for download route
Build download controller scaffolding using make:controller.

$ sail artisan make:controller --invokable -- DownloadController

   INFO  Controller [app/Http/Controllers/DownloadController.php] created successfully.

The __invoke() method is a magic method in PHP that allows an instance
of a class to be called as a function.  In Laravel, it is used in
controller classes for handling single-action logic.  When a controller
contains only the __invoke() method, it makes the class callable.

Build the __invoke() method with the closure of the download route that
now focuses on the route exclusively.