Use Blade echo construct for action buttons
Blade echo construct of {{ __(...) }} offers many benefits over text by
itself, such as preventing Cross-Site Scripting (XSS) attacks as well as
providing localization support.  It is preferable to use Blade echo
construct for action buttons in our views.

Balance the Cancel and Delete buttons with the same width and center the
text for visual appeal.

As we work on related Blade views, format them for consistency.