Add Download button
Build Download button to download a file from the app and place it next
to Create Job button.

Download link was in the navlinks, but that introduced an issue.  When
we click on the Download link, it doesn't become active.  Download link
was in the wrong place and it is removed.