Add timestamps to job_tag pivot table
By default Laravel's Eloquent doesn't automatically populate the
created_at and updated_at timestamps columns in the job_tag pivot table
when using the attach method on a many-to-many relationship.  Use
withTimestamps() method in the relationship definition in our Job and
Tag models to let Laravel automatically manage the timestamps for the
pivot table.